requirements.txt
alembic==1.18.4
beautifulsoup4==4.14.3
Flask==3.1.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
gunicorn
lxml
nltk==3.9.3
numpy==2.4.3
pandas==3.0.1
Requests==2.32.5
scikit_learn==1.4.2
scipy==1.17.1
SQLAlchemy==2.0.48
vaderSentiment==3.3.2
yfinance==1.2.0
psycopg2-binary
curl_cffispacy
